% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concs-data.R
\docType{data}
\name{concs}
\alias{concs}
\title{example of concentrations for a reaction data set}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(concs)
}
\description{
Reaction data set of a random bunch of reactions
}
\examples{
data(concs)
}
\keyword{datasets}
