% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CRNs-data.R
\docType{data}
\name{crns}
\alias{crns}
\title{example reaction set for a chemical reaction network}
\format{
An object of class \code{"list"}
}
\usage{
data(CRNs)
}
\description{
Reaction data set of a random bunch of reactions
}
\examples{
data(CRNs)
}
\keyword{datasets}
