% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crn.R
\name{make_concs_file}
\alias{make_concs_file}
\title{Make Species Concentrations from Chemical Network}
\usage{
make_concs_file(init_path, reactions)
}
\arguments{
\item{init_path}{The name of the file containing the concentrations of species}

\item{reactions}{A character vector containing strings with the reaction equations}
}
\value{
a file containing a template for species concentrations or a warning
}
\description{
Make Species Concentrations from Chemical Network
}
\examples{
data(CRNs)
make_concs_file(init_path = tempfile(), crns$rps$rxns)
}
