% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCRNs-package.R
\docType{package}
\name{RCRNs-package}
\alias{RCRNs}
\alias{RCRNs-package}
\title{RCRNs: Simulate Chemical Reaction Networks}
\description{
R package that provides tools for modeling chemical reaction networks.
}
\author{
\strong{Maintainer}: Kern Hast \email{Kernith@gmail.com} (\href{https://orcid.org/0000-0001-8770-8420}{ORCID})

}
\keyword{internal}
