% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rxns-data.R
\docType{data}
\name{rxns}
\alias{rxns}
\title{example reaction set for a chemical reaction network}
\format{
An object of class \code{"data.frame"}
}
\usage{
data(rxns)
}
\description{
Reaction data set of a random bunch of reactions
}
\examples{
data(rxns)
}
\keyword{datasets}
