% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crn.R
\name{simulate_crn}
\alias{simulate_crn}
\title{Simulate a Chemical Reaction Network}
\usage{
simulate_crn(
  rxn_params,
  init_concs,
  time_step = 10^-3,
  max_time = 60,
  output_res = 1000
)
}
\arguments{
\item{rxn_params}{A character vector containing strings with the reaction
equations. This set of parameters can be generated from a human-friendly
strings via the reaction equations with the \code{parse_reactions} function.}

\item{init_concs}{A data frame containing initial concentrations of chemical
species in the reaction network. Species not supplied with an initial
concentration are assumed to have an initial concentration of 0.}

\item{time_step}{The amount of time that elapses in each simulation cycle}

\item{max_time}{The full amount of time the simulation will run}

\item{output_res}{The number of time points that will be returned from the
simulation}
}
\value{
Returns a data frame with the concentrations and the reaction rates
at each output time point
}
\description{
Simulates the progression of a network of chemical reactions from a set of
initial concentrations.
}
\details{
Calculates reaction rates and changes in concentration for a chemical
reaction network described by \code{rxn_params}. Calculations are done using
classic mass-action rate equations, and are performed at a series of time
points incremented by \code{time_step} from 0 to \verb{max time} . \cr \cr The units of
measure (e.g. molar or seconds) for \code{time_step} and \code{max_time} and
\code{rxn_params} (rates, rate constants, and concentrations) are arbitrary, but
are assumed to be consistent (e.g. if concentration is supplied in nanomolar
and \code{time_step} in seconds, rate is assumed to be in nanomolar per second).
}
\examples{
data(CRNs)
results <- simulate_crn(
  rxn_params = parse_reactions(crns$rps$rxns),
  init_concs = crns$rps$concs
)

}
