% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crn.R
\name{parse_reactions}
\alias{parse_reactions}
\title{Parse a chemical reaction network}
\usage{
parse_reactions(rxn_path, partial_orders = NA)
}
\arguments{
\item{rxn_path}{The name of the file containing the reaction equation set}

\item{partial_orders}{The name of the file containing the partial orders}
}
\value{
A list of data frames with reaction parameters
}
\description{
Parse a chemical reaction network
}
\examples{
parse_reactions("rxns.txt")
}
