% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crn.R
\name{parse_reactions}
\alias{parse_reactions}
\title{Parse a chemical reaction network}
\usage{
parse_reactions(reactions, partial_orders = NULL)
}
\arguments{
\item{reactions}{A character vector containing strings with the reaction equations}

\item{partial_orders}{The name of the file containing the partial orders}
}
\value{
A list of data frames with reaction parameters
}
\description{
Parse a chemical reaction network
}
\examples{
data(CRNs)
parse_reactions(crns$rps$rxns)
}
