% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_crn.R
\name{simulate_reaction}
\alias{simulate_reaction}
\title{Title}
\usage{
simulate_reaction(
  reactions,
  init_concs,
  partial_orders = NULL,
  time_step = 10^-3,
  max_time = 60,
  output_res = 1000
)
}
\arguments{
\item{reactions}{A character vector containing strings with the reaction equations}

\item{init_concs}{The name of the file containing the concentrations of species}

\item{partial_orders}{The name of the file containing the partial orders}

\item{time_step}{the amount of time that elapses in each simulation cycle}

\item{max_time}{the full amount of time the simulation will run}

\item{output_res}{the number of time points that will be returned from the simulation}
}
\value{
a data frame with the concentrations and reaction rates at each output time point
}
\description{
Title
}
\examples{
data(rxns)
data(concs)
results <- simulate_reaction(
  reactions = rxns,
  init_concs = concs,
  time_step = 10^-2,
  max_time = 360,
  output_res = 1000
  )
}
